+++ apm-agent-core/src/main/java/co/elastic/apm/agent/impl/context/Db.java
@@ -23,7 +23,7 @@
  * #L%
  */
 
-package co.elastic.apm.agent.impl.transaction;
+package co.elastic.apm.agent.impl.context;
 
 import co.elastic.apm.agent.objectpool.Allocator;
 import co.elastic.apm.agent.objectpool.ObjectPool;