+++ apm-agent-core/src/main/java/co/elastic/apm/agent/impl/context/Http.java
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.impl.transaction;
+package co.elastic.apm.agent.impl.context;
 
 import co.elastic.apm.agent.objectpool.Recyclable;